requirements.txt
django
djangorestframework
psycopg2-binary
drf-yasg
gunicorn
psycopg2
